Add result icons and make remaining options move
The selected answer and the correct answer now move to the bottom
Added vectors for a green "tick" and a red "cross" inplace of the timer
when time runs out. Added an image view to display it.
Created references to the parent RelativeLayouts of all the options, and
set their visibility to GONE to prevent invisible options from taking
screen space.